Structure tests into multiple files
